agentic_etl_app_old/requirements.txt
streamlit==1.36.0
pandas==2.2.2
numpy==1.26.4
jinja2==3.1.4
boto3==1.34.131
snowflake-connector-python==3.11.0
redshift-connector==2.1.2
databricks-sql-connector==3.5.0
openpyxl==3.1.5
PyYAML==6.0.2
weasyprint==62.3    # optional; falls back to HTML if not working
